images

+++ app/sidekiq/delete_image_job.rb
@@ -1,10 +1,13 @@
 class DeleteImageJob
   include Sidekiq::Job
 
-  def perform(image_id)
-    image = Image.find(image_id)
-    image.destroy!
-    Rails.logger.info "Image #{image_id} deleted"
+  def perform(image_ids)
+    images = Image.where(id: image_ids)
+    return if images.empty?
+    images.each do |image|
+      image.destroy!
+      Rails.logger.info "Image #{image.id} deleted"
+    end
     # Do something
   end
 end
